+++ core/terraform/keyvault.tf
@@ -116,6 +116,7 @@ resource "azurerm_monitor_diagnostic_setting" "kv" {
   name                       = "diagnostics-kv-${var.tre_id}"
   target_resource_id         = azurerm_key_vault.kv.id
   log_analytics_workspace_id = module.azure_monitor.log_analytics_workspace_id
+  log_analytics_destination_type = "AzureDiagnostics"
 
   dynamic "log" {
     for_each = toset(["AuditEvent", "AzurePolicyEvaluationDetails"])