+++ templates/shared_services/firewall/terraform/locals.tf
@@ -11,4 +11,7 @@ locals {
     tre_id                = var.tre_id
     tre_shared_service_id = var.tre_resource_id
   }
+
+  api_driven_application_rule_collection = jsondecode(base64decode(var.api_driven_rule_collections_b64))
+  api_driven_network_rule_collection = jsondecode(base64decode(var.api_driven_network_rule_collections_b64))
 }